chore: Move Sentry and FileStore modules to modules/
These are modules! Why would they not go in the modules/ directory???? (thank god for webstorm)
