+++ server/src/modules/filestore/file-store.interface.ts
